@@ -26,37 +26,19 @@ class TagSubscriber extends AbstractConsumer implements ConsumerInterface
 {
 
     /**
-     * @var TagContactBuilder
-     */
-    private $requestBuilder;
-
-    /**
-     * @var Client
-     */
-    private $client;
-
-    /**
-     * @var ContactRepositoryInterface
-     */
-    private $contactRepository;
-
-    /**
      * AssignContactToListConsumer constructor.
      *
-     * @param TagContactBuilder          $tagContactBuilder
+     * @param TagContactBuilder $requestBuilder
      * @param Client                     $client
      * @param ContactRepositoryInterface $contactRepository
      */
     public function __construct(
-        TagContactBuilder $tagContactBuilder,
-        Client $client,
-        ContactRepositoryInterface $contactRepository,
+        private TagContactBuilder $requestBuilder,
+        private readonly Client $client,
+        private readonly ContactRepositoryInterface $contactRepository,
         Logger $logger
     ) {
         parent::__construct($logger);
-        $this->requestBuilder    = $tagContactBuilder;
-        $this->client            = $client;
-        $this->contactRepository = $contactRepository;
     }
 
     /**
@@ -64,7 +46,7 @@ class TagSubscriber extends AbstractConsumer implements ConsumerInterface
      */
     public function consume(string $message): void
     {
-        $message = json_decode($message, true);
+        $message = json_decode($message, true, 512, JSON_THROW_ON_ERROR);
         $contact = $this->contactRepository->getById($message['contact_id']);
 
         $tagIds   = $message['tags'];
@@ -84,8 +66,6 @@ class TagSubscriber extends AbstractConsumer implements ConsumerInterface
     }
 
     /**
-     * @param ContactInterface $contact
-     * @param array            $tagIds
      *
      * @return array
      */
@@ -93,9 +73,7 @@ class TagSubscriber extends AbstractConsumer implements ConsumerInterface
     {
         $requestBuilder = $this->requestBuilder;
         return array_map(
-            function ($tagId) use ($requestBuilder, $contact) {
-                return $requestBuilder->buildWithContact($contact, (int)$tagId);
-            },
+            fn($tagId) => $requestBuilder->buildWithContact($contact, (int)$tagId),
             $tagIds
         );
     }
