@@ -22,26 +22,8 @@ use Magento\Framework\MessageQueue\PublisherInterface;
 class AssignContactToListObserver implements ObserverInterface
 {
 
-    /**
-     * @var ConfigHelper
-     */
-    private $configHelper;
-
-    /**
-     * @var PublisherInterface
-     */
-    private $publisher;
-
-    /**
-     * @param ConfigHelper       $configHelper
-     * @param PublisherInterface $publisher
-     */
-    public function __construct(
-        ConfigHelper $configHelper,
-        PublisherInterface $publisher
-    ) {
-        $this->configHelper = $configHelper;
-        $this->publisher    = $publisher;
+    public function __construct(private readonly ConfigHelper $configHelper, private readonly PublisherInterface $publisher)
+    {
     }
 
     /**
@@ -64,7 +46,7 @@ class AssignContactToListObserver implements ObserverInterface
                 [
                     'contact_id' => $contact->getId(),
                     'list_id'    => $this->configHelper->getCustomerListId()
-                ]
+                ], JSON_THROW_ON_ERROR
             )
         );
     }
