@@ -17,26 +17,8 @@ use Magento\Framework\MessageQueue\PublisherInterface;
  */
 class ExportCustomerObserver implements ObserverInterface
 {
-    /**
-     * @var ConfigHelper
-     */
-    private $configHelper;
-
-    /**
-     * @var PublisherInterface
-     */
-    private $publisher;
-
-    /**
-     * @param ConfigHelper $configHelper
-     * @param PublisherInterface $publisher
-     */
-    public function __construct(
-        ConfigHelper $configHelper,
-        PublisherInterface $publisher
-    ) {
-        $this->configHelper = $configHelper;
-        $this->publisher = $publisher;
+    public function __construct(private readonly ConfigHelper $configHelper, private readonly PublisherInterface $publisher)
+    {
     }
 
     /**
@@ -53,7 +35,7 @@ class ExportCustomerObserver implements ObserverInterface
 
         $this->publisher->publish(
             Topics::CUSTOMER_CUSTOMER_EXPORT,
-            json_encode(['magento_customer_id' => $magentoCustomer->getId()])
+            json_encode(['magento_customer_id' => $magentoCustomer->getId()], JSON_THROW_ON_ERROR)
         );
     }
 }
