@@ -17,26 +17,8 @@ use CommerceLeague\ActiveCampaign\Helper\Config as ConfigHelper;
  */
 class ExportContactObserver implements ObserverInterface
 {
-    /**
-     * @var ConfigHelper
-     */
-    private $configHelper;
-
-    /**
-     * @var PublisherInterface
-     */
-    private $publisher;
-
-    /**
-     * @param ConfigHelper $configHelper
-     * @param PublisherInterface $publisher
-     */
-    public function __construct(
-        ConfigHelper $configHelper,
-        PublisherInterface $publisher
-    ) {
-        $this->configHelper = $configHelper;
-        $this->publisher = $publisher;
+    public function __construct(private readonly ConfigHelper $configHelper, private readonly PublisherInterface $publisher)
+    {
     }
 
     /**
@@ -57,7 +39,7 @@ class ExportContactObserver implements ObserverInterface
 
         $this->publisher->publish(
             Topics::NEWSLETTER_CONTACT_EXPORT,
-            json_encode(['email' => $subscriber->getEmail()])
+            json_encode(['email' => $subscriber->getEmail()], JSON_THROW_ON_ERROR)
         );
     }
 }
