@@ -18,34 +18,8 @@ use Magento\Framework\MessageQueue\PublisherInterface;
 class ExportOmittedGuestCustomers implements CronInterface
 {
 
-    /**
-     * @var ConfigHelper
-     */
-    private $configHelper;
-
-    /**
-     * @var CustomerCollectionFactory
-     */
-    private $customerCollectionFactory;
-
-    /**
-     * @var PublisherInterface
-     */
-    private $publisher;
-
-    /**
-     * @param ConfigHelper              $configHelper
-     * @param CustomerCollectionFactory $customerCollectionFactory
-     * @param PublisherInterface        $publisher
-     */
-    public function __construct(
-        ConfigHelper $configHelper,
-        CustomerCollectionFactory $customerCollectionFactory,
-        PublisherInterface $publisher
-    ) {
-        $this->configHelper              = $configHelper;
-        $this->customerCollectionFactory = $customerCollectionFactory;
-        $this->publisher                 = $publisher;
+    public function __construct(private readonly ConfigHelper $configHelper, private readonly CustomerCollectionFactory $customerCollectionFactory, private readonly PublisherInterface $publisher)
+    {
     }
 
     /**
