@@ -23,27 +23,10 @@ class TagNewsletterSubscriberObserver implements ObserverInterface
 {
 
     /**
-     * @var Config
-     */
-    private $configHelper;
-
-    /**
-     * @var PublisherInterface
-     */
-    private $publisher;
-
-    /**
      * TagNewsletterSubscriberObserver constructor.
-     *
-     * @param Config             $configHelper
-     * @param PublisherInterface $publisher
      */
-    public function __construct(
-        Config $configHelper,
-        PublisherInterface $publisher
-    ) {
-        $this->configHelper = $configHelper;
-        $this->publisher    = $publisher;
+    public function __construct(private readonly Config $configHelper, private readonly PublisherInterface $publisher)
+    {
     }
 
     /**
@@ -64,7 +47,7 @@ class TagNewsletterSubscriberObserver implements ObserverInterface
         if (null !== $tags) {
             $this->publisher->publish(
                 Topics::TAG_NEWSLETTER_SUBSCRIBER,
-                json_encode(['contact_id' => $contact->getId(), 'tags' => $tags])
+                json_encode(['contact_id' => $contact->getId(), 'tags' => $tags], JSON_THROW_ON_ERROR)
             );
         }
     }
